refactor: make SideNav accordions work with CloseButton
